refactor cloudwatch filenames
