+++ javascriptv3/example_code/cloudwatch/src/cloudwatch-disableAlarmActions.js
@@ -18,7 +18,7 @@ node cw_disablealarmactions.js
 
 // Import required AWS SDK clients and commands for Node.js
 import { DisableAlarmActionsCommand } from "@aws-sdk/client-cloudwatch";
-import { cwClient } from "./libs/cwClient";
+import { cwClient } from "./libs/cloudwatch-cloudWatchClient";
 
 // Set the parameters
 const params = { AlarmNames: "ALARM_NAME" }; // e.g., "Web_Server_CPU_Utilization"
