+++ javascriptv3/example_code/cloudwatch/src/cloudwatch-events-putRule.js
@@ -18,7 +18,7 @@ node cw_deletealarm.js
 
 // Import required AWS SDK clients and commands for Node.js
 import { PutRuleCommand } from "@aws-sdk/client-cloudwatch-events";
-import { cweClient } from "./libs/cweClient";
+import { cweClient } from "./libs/cloudwatch-events-cloudWatchEventsClient";
 
 // Set the parameters
 const params = {
