+++ javascriptv3/example_code/cloudwatch/src/libs/cloudwatch-events-cloudWatchEventsClient.js
