+++ javascriptv3/example_code/cloudwatch/src/libs/cloudwatch-logs-cloudwatchLogsClient.js
